--- src/main/java/sopt/org/HMH/common/exception/base/RootEnum.java
@@ -1,4 +1,0 @@
-package sopt.org.HMH.common.exception.base;
-
-public interface RootEnum {
-}
