+++ src/main/java/sopt/org/HMH/global/common/exception/GlobalException.java
@@ -1,6 +1,6 @@
-package sopt.org.HMH.common.exception;
+package sopt.org.HMH.global.common.exception;
 
-import sopt.org.HMH.common.exception.base.ExceptionBase;
+import sopt.org.HMH.global.common.exception.base.ExceptionBase;
 
 /**
  * 전역적 혹은 시스템 전체 범위에서 발생하는 예외
