+++ src/main/java/sopt/org/HMH/global/common/exception/base/RootEnum.java
@@ -1,0 +1,4 @@
+package sopt.org.HMH.global.common.exception.base;
+
+public interface RootEnum {
+}
